+++ src/KubernetesPfSenseController/Plugin/MetalLB.php
@@ -28,9 +28,12 @@ class MetalLB extends PfSenseAbstract
         $pluginConfig = $this->getConfig();
         $nodeLabelSelector = $pluginConfig['nodeLabelSelector'];
         $nodeFieldSelector = $pluginConfig['nodeFieldSelector'];
+        $configMap = ($pluginConfig['configMap']) ? $pluginConfig['configMap'] : "metallb-system/config";
+        $configMapNamespace = explode("/", $configMap)[0];
+        $configMapName = explode("/", $configMap)[1];
 
         // metallb config
-        $watch = $controller->getKubernetesClient()->createWatch('/api/v1/watch/namespaces/metallb-system/configmaps/config', [], $this->getMetalLbConfigWatchCallback());
+        $watch = $controller->getKubernetesClient()->createWatch("/api/v1/watch/namespaces/${configMapNamespace}/configmaps/${configMapName}", [], $this->getMetalLbConfigWatchCallback());
         $this->addWatch($watch);
 
         // initial load of nodes
@@ -38,7 +41,7 @@ class MetalLB extends PfSenseAbstract
             'labelSelector' => $nodeLabelSelector,
             'fieldSelector' => $nodeFieldSelector,
         ];
-        $nodes = $controller->getKubernetesClient()->request('/api/v1/nodes', 'GET', $params);
+        $nodes = $controller->getKubernetesClient()->createList('/api/v1/nodes', $params)->get();
         $this->state['nodes'] = $nodes['items'];
 
         // watch for node changes
@@ -111,7 +114,8 @@ class MetalLB extends PfSenseAbstract
 
         switch ($pluginConfig['bgp-implementation']) {
             case 'openbgp':
-                return $this->doActionOpenbgp();
+            case 'frr':
+                return $this->doActionGeneric();
                 break;
             default:
                 $this->log('unsupported bgp-implementation: '.$pluginConfig['bgp-implementation']);
@@ -121,23 +125,35 @@ class MetalLB extends PfSenseAbstract
     }
 
     /**
-     * Update pfSense state for openbgp
+     * Update pfSense state for bgp implementation
      *
      * @return bool
      */
-    private function doActionOpenbgp()
+    private function doActionGeneric()
     {
         $metalConfig = $this->state['metallb-config'];
-        $openbgpConfig = PfSenseConfigBlock::getInstalledPackagesConfigBlock($this->getController()->getRegistryItem('pfSenseClient'), 'openbgpdneighbors');
-        if (empty($openbgpConfig->data)) {
-            $openbgpConfig->data = [];
-        }
-
-        if (empty($openbgpConfig->data['config'])) {
-            $openbgpConfig->data['config'] = [];
-        }
-
         $pluginConfig = $this->getConfig();
+
+        switch ($pluginConfig['bgp-implementation']) {
+            case 'openbgp':
+                $bgpConfig = PfSenseConfigBlock::getInstalledPackagesConfigBlock($this->getController()->getRegistryItem('pfSenseClient'), 'openbgpdneighbors');
+                break;
+            case 'frr':
+                $bgpConfig = PfSenseConfigBlock::getInstalledPackagesConfigBlock($this->getController()->getRegistryItem('pfSenseClient'), 'frrbgpneighbors');
+                break;
+            default:
+                $this->log('unsupported bgp-implementation: '.$pluginConfig['bgp-implementation']);
+                return false;
+                break;
+        }
+
+        if (empty($bgpConfig->data)) {
+            $bgpConfig->data = [];
+        }
+
+        if (empty($bgpConfig->data['config'])) {
+            $bgpConfig->data['config'] = [];
+        }
 
         $bgpEnabled = false;
         foreach ($metalConfig['address-pools'] as $pool) {
@@ -149,7 +165,29 @@ class MetalLB extends PfSenseAbstract
 
         if ($bgpEnabled) {
             // add/remove as necessary
-            $template = $pluginConfig['options']['openbgp']['template'];
+            $template = $pluginConfig['options'][$pluginConfig['bgp-implementation']]['template'];
+            switch ($pluginConfig['bgp-implementation']) {
+                case 'openbgp':
+                    $defaults = [
+                        'md5sigkey' => (string) $template['md5sigkey'],
+                        'md5sigpass' => (string) $template['md5sigpass'],
+                        'groupname' => (string) $template['groupname'],
+                    ];
+                    $template = array_merge($defaults, $template);
+                    $template = array_map(function($v){
+                        return $v ?: '';
+                    },$template);
+                    break;
+                case 'frr':
+                    $defaults = [
+                        "sendcommunity" => "disabled",
+                    ];
+                    $template = array_merge($defaults, $template);
+                    $template = array_map(function($v){
+                        return $v ?: '';
+                    },$template);
+                    break;
+            }
 
             $nodes = $this->state['nodes'];
             $neighbors = [];
@@ -159,46 +197,58 @@ class MetalLB extends PfSenseAbstract
                 $managedNeighborsPreSave[$host] = [
                     'resource' => $this->getKubernetesResourceDetails($node),
                 ];
-                $neighbor = [
-                    'descr' => $host,
-                    'neighbor' => KubernetesUtils::getNodeIp($node),
-                    'md5sigkey' => (string) $template['md5sigkey'],
-                    'md5sigpass' => (string) $template['md5sigpass'],
-                    'groupname' => (string) $template['groupname'],
-                    'row' => $template['row'],
-                ];
+                $neighbor = $template;
+
+                switch ($pluginConfig['bgp-implementation']) {
+                    case 'openbgp':
+                        $neighbor['descr'] = $host;
+                        $neighbor['neighbor'] = KubernetesUtils::getNodeIp($node);
+                        break;
+                    case 'frr':
+                        $neighbor['descr'] = $host;
+                        $neighbor['peer'] = KubernetesUtils::getNodeIp($node);
+                        break;
+                }
+
                 $neighbors[] = $neighbor;
             }
 
             // get store data
             $store = $this->getStore();
             $managedNeighborNamesPreSave = @array_keys($managedNeighborsPreSave);
-            $managedNeighborNames = @array_keys($store['openbgp']['managed_neighbors']);
+            $managedNeighborNames = @array_keys($store[$pluginConfig['bgp-implementation']]['managed_neighbors']);
             if (empty($managedNeighborNames)) {
                 $managedNeighborNames = [];
             }
 
             // update config with new/updated items
             foreach ($neighbors as $neighbor) {
-                Utils::putListItem($openbgpConfig->data['config'], $neighbor, 'descr');
+                Utils::putListItem($bgpConfig->data['config'], $neighbor, 'descr');
             }
 
             // remove items from config
             $toDeleteItemNames = array_diff($managedNeighborNames, $managedNeighborNamesPreSave);
             foreach ($toDeleteItemNames as $itemId) {
-                Utils::removeListItem($openbgpConfig->data['config'], $itemId, 'descr');
+                Utils::removeListItem($bgpConfig->data['config'], $itemId, 'descr');
             }
 
             // prep config for save
-            if (empty($openbgpConfig->data['config'])) {
-                $openbgpConfig->data = null;
+            if (empty($bgpConfig->data['config'])) {
+                $bgpConfig->data = null;
             }
 
             // save newly managed configuration
             try {
-                $this->savePfSenseConfigBlock($openbgpConfig);
-                $this->reloadOpenbgp();
-                $store['openbgp']['managed_neighbors'] = $managedNeighborsPreSave;
+                $this->savePfSenseConfigBlock($bgpConfig);
+                switch ($pluginConfig['bgp-implementation']) {
+                    case 'openbgp':
+                        $this->reloadOpenbgp();
+                        break;
+                    case 'frr':
+                        $this->reloadFrrBgp();
+                        break;
+                }
+                $store[$pluginConfig['bgp-implementation']]['managed_neighbors'] = $managedNeighborsPreSave;
                 $this->saveStore($store);
 
                 return true;
@@ -210,20 +260,27 @@ class MetalLB extends PfSenseAbstract
             //remove any nodes from config
             // get storage data
             $store = $this->getStore();
-            $managedNeighborNames = @array_keys($store['openbgp']['managed_neighbors']);
+            $managedNeighborNames = @array_keys($store[$pluginConfig['bgp-implementation']]['managed_neighbors']);
             if (empty($managedNeighborNames)) {
                 return true;
             }
 
             foreach ($managedNeighborNames as $itemId) {
-                Utils::removeListItem($openbgpConfig->data['config'], $itemId, 'descr');
+                Utils::removeListItem($bgpConfig->data['config'], $itemId, 'descr');
             }
 
             // save newly managed configuration
             try {
-                $this->savePfSenseConfigBlock($openbgpConfig);
-                $this->reloadOpenbgp();
-                $store['openbgp']['managed_neighbors'] = [];
+                $this->savePfSenseConfigBlock($bgpConfig);
+                switch ($pluginConfig['bgp-implementation']) {
+                    case 'openbgp':
+                        $this->reloadOpenbgp();
+                        break;
+                    case 'frr':
+                        $this->reloadFrrBgp();
+                        break;
+                }
+                $store[$pluginConfig['bgp-implementation']]['managed_neighbors'] = [];
                 $this->saveStore($store);
 
                 return true;